re organise api
