+++ packages/facility-server/app/routes/apiv1/patient/patient.js
@@ -19,7 +19,10 @@ import { patientInvoiceRoutes } from './patientInvoice';
 import { patientRelations } from './patientRelations';
 import { patientBirthData } from './patientBirthData';
 import { patientLocations } from './patientLocations';
-import { patientProgramRegistration } from './patientProgramRegistration';
+import {
+  patientProgramRegistration,
+  patientProgramRegistrationConditions,
+} from './patientProgramRegistration';
 import { getOrderClause } from '../../../database/utils';
 import { dbRecordToResponse, pickPatientBirthData, requestBodyToRecord } from './utils';
 import { PATIENT_SORT_KEYS } from './constants';
@@ -222,7 +225,7 @@ patientRoute.get(
     });
 
     const { count } = lastEncounterMedications;
-    const data = lastEncounterMedications.rows.map(x => x.forResponse());
+    const data = lastEncounterMedications.rows.map((x) => x.forResponse());
 
     res.send({
       count,
@@ -253,14 +256,14 @@ patientRoute.get(
       PATIENT_SORT_KEYS.firstName,
       PATIENT_SORT_KEYS.displayId,
     ]
-      .filter(v => v !== orderBy)
-      .map(v => `${v} ASC`)
+      .filter((v) => v !== orderBy)
+      .map((v) => `${v} ASC`)
       .join(', ');
 
     // query is always going to come in as strings, has to be set manually
     ['ageMax', 'ageMin']
-      .filter(k => filterParams[k])
-      .forEach(k => {
+      .filter((k) => filterParams[k])
+      .forEach((k) => {
         filterParams[k] = parseFloat(filterParams[k]);
       });
 
@@ -277,7 +280,7 @@ patientRoute.get(
     // 2.d) the same rule of 2.b is applied in case we have two or more columns starting with what the user selected.
     // 2.e) The last rule for selected filters, is, if the user has selected any of those filters, we should also sort them alphabetically.
     if (!orderBy) {
-      const selectedFilters = ['displayId', 'lastName', 'firstName'].filter(v => filterParams[v]);
+      const selectedFilters = ['displayId', 'lastName', 'firstName'].filter((v) => filterParams[v]);
       if (selectedFilters?.length) {
         filterSortReplacements = selectedFilters.reduce((acc, filter) => {
           return {
@@ -290,18 +293,20 @@ patientRoute.get(
         // Exact match sort
         const exactMatchSort = selectedFilters
           .map(
-            filter => `upper(patients.${snakeCase(filter)}) = ${`:exactMatchSort${filter}`} DESC`,
+            (filter) => `upper(patients.${snakeCase(filter)}) = ${`:exactMatchSort${filter}`} DESC`,
           )
           .join(', ');
 
         // Begins with sort
         const beginsWithSort = selectedFilters
-          .map(filter => `upper(patients.${snakeCase(filter)}) LIKE :beginsWithSort${filter} DESC`)
+          .map(
+            (filter) => `upper(patients.${snakeCase(filter)}) LIKE :beginsWithSort${filter} DESC`,
+          )
           .join(', ');
 
         // the last one is
         const alphabeticSort = selectedFilters
-          .map(filter => `patients.${snakeCase(filter)} ASC`)
+          .map((filter) => `patients.${snakeCase(filter)} ASC`)
           .join(', ');
 
         filterSort = `${exactMatchSort}, ${beginsWithSort}, ${alphabeticSort}`;
@@ -437,8 +442,9 @@ patientRoute.get(
         ${select}
         ${from}
 
-        ORDER BY  ${filterSort &&
-          `${filterSort},`} ${sortKey} ${sortDirection}, ${secondarySearchTerm} NULLS LAST
+        ORDER BY  ${
+          filterSort && `${filterSort},`
+        } ${sortKey} ${sortDirection}, ${secondarySearchTerm} NULLS LAST
         LIMIT :limit
         OFFSET :offset
       `,
@@ -455,7 +461,7 @@ patientRoute.get(
       },
     );
 
-    const forResponse = result.map(x => renameObjectKeys(x.forResponse()));
+    const forResponse = result.map((x) => renameObjectKeys(x.forResponse()));
 
     res.send({
       data: forResponse,
@@ -515,6 +521,7 @@ patientRoute.use(patientInvoiceRoutes);
 patientRoute.use(patientBirthData);
 patientRoute.use(patientLocations);
 patientRoute.use(patientProgramRegistration);
+patientRoute.use(patientProgramRegistrationConditions);
 patientRoute.use(patientContact);
 
 export { patientRoute as patient };
