+++ packages/facility-server/app/routes/apiv1/patient/patientprogramRegistration/index.js
@@ -1,0 +1,2 @@
+export { patientProgramRegistrationConditions } from './patientProgramRegistrationConditions';
+export { patientProgramRegistration } from './patientProgramRegistration';
