+++ packages/web/app/features/ProgramRegistry/index.js
@@ -1,3 +1,4 @@
 export { ProgramRegistryConditionCategoryField } from './ProgramRegistryConditionCategoryField';
 export { ProgramRegistryConditionField } from './ProgramRegistryConditionField';
 export { UpdateConditionFormModal } from './UpdateConditionFormModal';
+export { PatientProgramRegistryUpdateFormModal } from './PatientProgramRegistryUpdateFormModal.jsx';
